Done alot of refactoring, and now made it so sprites can be drawn to the screen. In its current state the engine renders 5000 sprites from a single spriteatlas, bouncing around the screen. Whats currently lacking in spriterendering is to make sprites respond to a modelmatrix, worldmatrix and viewmatrix, enable rendering of sprites from multiple sprite atlases, and to optimize it further, with a goal of hitting about 50k sprites with acceptable frames per second.

+++ Abysswalker/scenes/Scene.cpp
@@ -19,7 +19,7 @@ void Scene::addEntity(Entity* e)
 
 void Scene::addEntities(std::vector<Entity*>& entities)
 {
-	for (int i = 0; i < entities.size(); i++) {
+	for (size_t i = 0; i < entities.size(); i++) {
 		addEntity(entities.back());
 		entities.pop_back();
 	}
